wip - refactoring of series AboutTab, still not done
